Post_Grouting/requirements.txt
Markdown==2.6.6

django
mysqlclient
pillow
awesome-slugify
sorl-thumbnail